requirements.txt
pytz==2022.2.1  # https://github.com/stub42/pytz
python-docx==0.8.11
docxtpl==0.16.4
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.3.0  # https://github.com/celery/django-celery-beat
flower==1.1.0  # https://github.com/mher/flower
uvicorn[standard]==0.18.2  # https://github.com/encode/uvicorn
pydantic==1.10.2 # https://github.com/pydantic/pydantic
requests==2.28.1 # https://github.com/psf/requests
channels==3.0.5 # https://github.com/django/channels
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
num2words==0.5.12

# Django
# ------------------------------------------------------------------------------
django==3.2.15  # pyup: < 4.0  # https://www.djangoproject.com/
django-filter==22.1
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.2.1
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
# REST Swagger API
django-rest-swagger==2.2.0
# DRF-spectacular for api documentation
drf-spectacular==0.23.1  # https://github.com/tfranzel/drf-spectacular